+++ test/helper.rb
@@ -7,7 +7,8 @@ require "minitest/autorun"
 require "active_record"
 require "byebug"
 
-require_relative "./test"
+require_relative "./case"
 
+ActiveRecord::Migration.verbose = false # Silence migrations
 ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: ":memory:")
 ActiveRecord::Base.logger = Logger.new($stdout) if ENV.fetch("LOG", false)
